requirements.txt
aiohttp==3.6.2
appdirs==1.4.3
asgiref==3.2.3
async-timeout==3.0.1
attrs==19.3.0
autoflake==1.3.1
autopep8==1.4.4
black==19.10b0
certifi==2019.11.28
chardet==3.0.4
Click==7.0
Django==3.0.1
django-crispy-forms==1.8.1
idna==2.8
isort==4.3.21
multidict==4.7.2
numpy==1.17.4
oauthlib==3.1.0
pandas==0.25.3
pathspec==0.6.0
Pillow==6.2.1
plotly==4.4.1
pycodestyle==2.5.0
pyflakes==2.1.1
python-dateutil==2.8.1
pytz==2019.3
regex==2019.12.20
requests==2.22.0
requests-oauthlib==1.3.0
retrying==1.3.3
six==1.13.0
sqlparse==0.3.0
toml==0.10.0
torch==1.5.1
torchvision==0.6.1
typed-ast==1.4.0
urllib3==1.25.7
yarl==1.4.2
